requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
rasterio==1.3.9
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.2
numpy==1.26.2
aiofiles==23.2.1
